Fix X509_REQ_INFO enc_len with Linux 64 bits
In OpenSSL sources, ASN1_ENCODING len is defined as a long, which is 32 bits on Windows 64 bits; but 64 bits on other systems (like GNU/Linux). Changing len type from int to IntPtr allows to handle both cases, and prevent data truncation.

+++ ManagedOpenSsl/X509/X509Request.cs
@@ -92,7 +92,7 @@ namespace OpenSSL.X509
 			#region ASN1_ENCODING enc;
 
 			public IntPtr enc_enc;
-			public int enc_len;
+			public IntPtr enc_len;
 			public int enc_modified;
 
 			#endregion
